% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexwall.R
\name{hexwall}
\alias{hexwall}
\title{hexwall}
\usage{
hexwall(path, sticker_row_size = 10, sticker_width = 200,
  remove_small = TRUE, remove_size = TRUE)
}
\arguments{
\item{path}{path to a folder of hexagon stickers}

\item{sticker_row_size}{number of stickers in the longest row}

\item{sticker_width}{width of each sticker (in pixels)}

\item{remove_small}{should hexagons smaller than sticker_width be removed?}

\item{remove_size}{Should hexagons of an abnormal size be removed?}
}
\description{
Neatly align hexagon stickers in R
}
\examples{
hexwall("samplehex", sticker_row_size = 3, sticker_width = 200)
}
